Don't create indexes on write
We're seeing a bug with TokuMX that creates a database-level lock on
ensure_index which is being triggered on every GridFS write. For now,
remove the calls to ensure_index in Mongo::Grid crudely. This version of
the driver is no longer supported and we should move to 2.x which does
not appear to create the index as aggressively.